# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Workcraft API
  description: |
    Studio-ops API for Workcraft. Response envelope: `{ status: "success"|"error", data? | message? }`.
    Authenticated routes expect `Authorization: Bearer <JWT>`. Domain routes require studio membership on the token.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiSuccess:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: success
        data:
          description: Endpoint-specific payload
...
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    PageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeQuery:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SearchQuery:
      name: search
      in: query
      schema:
        type: string
paths:
...
  /api/clients:
    get:
      tags: [Clients]
      summary: List clients
      operationId: listClients
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - $ref: "#/components/parameters/SearchQuery"
        - name: includeRetired
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Paginated clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Clients]
      summary: Create client
      operationId: createClient
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                phone:
                  type: string
                  nullable: true
                email:
                  type: string
                  nullable: true
                nricOrSsm:
                  type: string
                  nullable: true
                tin:
                  type: string
                  nullable: true
                socialHandle:
                  type: string
                  nullable: true
                address:
                  type: string
                  nullable: true
                whatsappConsent:
                  type: boolean
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/clients/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Clients]
      summary: Get client
      operationId: getClient
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Clients]
      summary: Update client
      operationId: updateClient
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Clients]
      summary: Soft-delete (retire) client
      operationId: deleteClient
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Retired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
  /api/account/profile:
    get:
      tags: [Account]
      summary: Get own profile
      operationId: getAccountProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile (user, studio, access)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Account]
      summary: Update own profile
      operationId: patchAccountProfile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                  nullable: true
                waUpdatesOptIn:
                  type: boolean
                myBillsRemindersEnabled:
                  type: boolean
                myBillsDaysBefore:
                  type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      operationId: listJobs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/PageSizeQuery"
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [CONFIRMED, COMPLETED, CANCELLED]
      responses:
        "200":
          description: Paginated jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    post:
      tags: [Jobs]
      summary: Create job
      operationId: createJob
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId]
              properties:
                clientId:
                  type: integer
                quotationId:
                  type: integer
                  nullable: true
                status:
                  type: string
                  enum: [CONFIRMED, COMPLETED, CANCELLED]
                sessions:
                  type: array
                  items:
                    type: object
                checklistItems:
                  type: array
                  items:
                    type: object
                deliverables:
                  type: array
                  items:
                    type: object
                contracts:
                  type: array
                  items:
                    type: object
      responses:
        "201":
          description: Created (number JOB-YYYY-####)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
  /api/jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Jobs]
      summary: Get job detail
      operationId: getJob
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Nested job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    patch:
      tags: [Jobs]
      summary: Update job
      operationId: updateJob
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
    delete:
      tags: [Jobs]
      summary: Delete job
      operationId: deleteJob
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...
  /api/calendar/sessions:
    get:
      tags: [Calendar]
      summary: List sessions in date range
      operationId: listCalendarSessions
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Sessions with job, client, crew
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiSuccess"
...